[wip] react my widgets graphs working
